userspace-dp: split wg_control.rs into wg_control/{mtu,sock,attempt,dispatch}.rs (#6438)
wg_control.rs was exactly 1,600 LOC of WG datapath control thread with
god-functions on the per-packet path: run_wg_control_loop (325 lines /
9 params), dispatch_inbound (209), and drive_attempt_machine (130),
fusing five layers — MTU math, the poll(2) wait, the handshake attempt
machine, the socket+cmsg codec, and inbound dispatch + encap. A module
that mixes those silently degrades reviewability and inlining, so split
it before the next change lands on it (docs/engineering-style.md
modularity discipline).

Pure code-motion within the crate:

- wg_control/mod.rs keeps the thread entry (wg_control_loop) and the
  orchestrating run_wg_control_loop (socket RX burst -> TUN-read burst
  -> 1s timer arm -> idle poll), the loop-level constants, and the
  submodule wiring.
- wg_control/mtu.rs holds WG_DEFAULT_OUTER_MTU, wg_encapped_size, and
  wg_inner_fits_outer_mtu.
- wg_control/sock.rs holds the socket+cmsg codec (bind_wg_socket,
  bind_dual_stack_v6, set_recv_tos_options, canonicalize_endpoint,
  wg_send_to, WgRecv, the align(8) CmsgBuf with its #2334 const assert,
  wg_recvmsg, parse_outer_ecn_from_cmsg,
  sockaddr_storage_to_socketaddr) plus the poll layer (PollWait,
  wg_poll_wait, poll_timeout_ms, WG_POLL_CAP_MS).
- wg_control/attempt.rs holds the #1888 S5 attempt state machine
  (HandshakeAttempt, AttemptTrigger, start_attempt,
  drive_attempt_machine, drive_initiation) and the keepalive emit/pace
  helpers (send_keepalive, pace_keepalive_skip).
- wg_control/dispatch.rs holds the InboundOutcome auth-before-roam
  contract, dispatch_inbound, and encap_and_send.
- wg_control_tests.rs (822 LOC) moves byte-identical into the
  directory (git rename); its use super::* binds the same names via
  mod.rs re-imports, so the suite passes unmodified.

Visibility is preserved exactly: moved items become pub(super) inside
their submodule — the same reach they had as file-private items.
WG_DEFAULT_OUTER_MTU keeps coordinator-tree visibility via a
pub(super) use at the historical wg_control::WG_DEFAULT_OUTER_MTU path
(tunnel_supervision.rs names it); the constant itself is declared
pub(in crate::afxdp::coordinator) in mtu.rs. Test-only seams
(wg_send_to, CmsgBuf, parse_outer_ecn_from_cmsg,
sockaddr_storage_to_socketaddr, wg_encapped_size,
wg_inner_fits_outer_mtu, WG_POLL_CAP_MS) re-import into mod.rs
#[cfg(test)]-gated, matching the #6436 gating precedent, so non-test
builds carry no unused imports. The four 64 KiB scratch buffers stay
once-allocated in the loop; #[inline] and
#[allow(clippy::too_many_arguments)] travel with their fns; fn-local
use statements move verbatim.

Also folds the pad_to_16 triplication (wg_control.rs, wg/engine.rs,
frame/wg.rs — three identical private const fns) into a single
pub(crate) const fn in afxdp/wg/mod.rs next to the other wire-format
constants; the three sites import it. The engine's const pad-proof
assert is import-path independent and keeps working.

Validation: cargo check / cargo check --tests warning parity vs the
4bc33a3b0 base (identical sorted warning sets, zero new warnings);
full crate suite cargo test --release --bins --tests --
--test-threads=1 green (4,290 passed / 0 failed / 2 ignored);
wg_control::tests 22/22 pass unmodified. asm-diff (objdump -Cd,
address/symbol normalized) on the release binary: wg_encap_frame
576/576 and WgEngine::encap_inner 278/278 identical modulo relocation
addresses (the pad fold is codegen-invisible), drive_initiation
160/160 identical, send_keepalive 186->176 (same stream modulo one
LLVM-eliminated dead 2-byte store, tail alignment padding, and the
resulting branch-offset shifts), wg_control_loop 2799->2781 with an
identical direct-call symbol multiset — deltas are register
allocation / stack offsets / NOP forms (CGU-reshuffle inlining
variance), no new arithmetic or control flow.

Docs: coordinator/README.md gains the wg_control/ file-table row;
stale wg_control.rs path citations in slowpath.rs, afxdp/gre.rs,
afxdp/wg/timers.rs, afxdp/wg/mod.rs, afxdp/frame/wg.rs, and
tunnel_supervision.rs updated. Plan: docs/pr/6438-wg-control-split/.
