+++ userspace-dp/src/afxdp/coordinator/README.md
@@ -25,7 +25,8 @@ the workers share.
 | `snapshot_refresh.rs` | `refresh_runtime_snapshot{,_disarmed,_inner}` (armed/disarmed same-plan snapshot-apply legs: preflight, #1873 tunnel-remap purge, forwarding swap + stores, aux-thread reconcile, CoS owner-map + warm passes) and `refresh_fabric_links`. (#1890 split.) **#5166: the CoS owner/lease maps + `ha.fabrics` are published BEFORE the `ha.forwarding` worker-visible store — a live worker (loop_body loads forwarding then CoS in one tick) must never observe new queue config without its matching CoS owners/leases (else transient class blackhole / stale-rate / over-admission).** **#3766: `_inner` is a FALLIBLE atomic swap — it returns `Result<(), SnapshotIntegrityError>`, builds the new forwarding state FIRST, and only then bumps `self.validation` + rotates the neighbor-manager keys + publishes; on a build integrity error nothing mutates and the error is returned so the handler fails closed (no split-brain, no persisted reject).** |
 | `status.rs` | Read-side snapshots for `show ...` queries. **#5289: `recent_exceptions()` / `last_resolution()` drain the per-worker POD exception rings (#6242: one `Arc<Mutex<ExceptionEventRing>>` per worker, now on each `WorkerRuntimeRecord` in `workers.records` — was the standalone `Coordinator::worker_exception_rings` / `worker_last_resolution` maps, mirroring the also-consolidated `worker_panics`) plus the control-thread ring, format each compact `ExceptionEvent`/`ResolutionEvent` into the operator-visible `ExceptionStatus`/`PacketResolution` HERE (interface/reason/IP/zone strings + monotonic→wall-clock via a single captured `MonoWallAnchor`), and merge by timestamp. The retired inline path formatted + locked a process-global mutex on EVERY terminal packet — the cross-worker DoS closed by #5289. The record path (`disposition::record_exception`) now writes a `&'static`-reason / `Arc<str>`-interface / `Copy`-`IpAddr` / numeric-zone POD event under a per-worker mutex with a per-(reason,5-tuple) sampler; the batched `BindingLiveState` counters stay the lossless signal. #6101: the slow-path reinject-failure sites (`_rate_limited` / `_queue_full` / `_slow_path_mtu_exceeded` / `_enqueue_failed`) record via `disposition::record_exception_suffixed`, which stores a `&'static` base reason + a `&'static` suffix (new `ExceptionEvent::reason_suffix`) so the record path stays alloc-free under a reinject-failure flood — the operator text `"{reason}{suffix}"` is reconstructed byte-identically by `ExceptionEvent::reason()` on the status thread. `record_exception_owned` remains ONLY for genuinely dynamic reasons (the `cfg!(feature = "debug-log")` forward-tuple-mismatch diagnostic). The cross-worker merge/sort/cap (`recent_exceptions`/`last_resolution` across ≥2 per-worker rings + the control ring) and the bring-up insert / teardown record-drop are unit-tested in `status_tests.rs::exception_ring_merge_6101`.** The other read-side exception is `drain_session_deltas`, which mutates per-binding state. **#5290: this RPC-fallback drain is FAIR — it spreads a `budget/num_bindings` quantum across live bindings via a rotating cursor (`WorkerManager::session_delta_drain_cursor`) using the shared `session_delta::drain_session_deltas_fair` helper, instead of handing the whole budget to the first slot with an early break. A low-slot worker can no longer consume the caller-wide budget and starve higher slots during the event-stream fallback. On budget overflow (undrained deltas remain) it arms `BindingLiveState::set_delta_loss` on the residual bindings; the owning worker loop folds that latch into `SessionTable::set_delta_loss` for the existing #2442 owner-RG resync, so no delta is silently lost. The owner-RG bulk-export mirror `ha.rs::drain_session_deltas_from_live` shares the same fair helper but does NOT arm the latch (it IS the resync).** |
 | `supervisor.rs` | `spawn_supervised_worker` / `spawn_supervised_aux` — catches panics, marks the worker dead on its `WorkerRuntimeAtomics`, captures a panic message into a per-worker slot. (#925 Phase 1.) |
-| `tunnel_supervision.rs` | GRE local-origin + WG control-thread LIFECYCLE (three-pass reconcile, tombstone backoff, periodic liveness sweeps, defer-branch snapshot prunes — see "Aux tunnel threads" below). The thread bodies live in `wg_control.rs` / `afxdp/tunnel.rs`; the entry maps (`tunnel_sources`, `wg_control_threads`) stay on `Coordinator` in `mod.rs`. (#1890 split.) |
+| `tunnel_supervision.rs` | GRE local-origin + WG control-thread LIFECYCLE (three-pass reconcile, tombstone backoff, periodic liveness sweeps, defer-branch snapshot prunes — see "Aux tunnel threads" below). The thread bodies live in `wg_control/` / `afxdp/tunnel.rs`; the entry maps (`tunnel_sources`, `wg_control_threads`) stay on `Coordinator` in `mod.rs`. (#1890 split.) |
+| `wg_control/` | WG datapath control-thread BODY (one supervised aux thread per `mode == "wireguard"` endpoint, kernel UDP socket + wgN TUN). `mod.rs` holds the thread entry (`wg_control_loop`) and the orchestrating `run_wg_control_loop` (socket RX burst → TUN-read burst → 1s timer arm → idle poll(2)); the fused layers are split into `mtu.rs` (pad-aware encapped-size formula + per-peer outer-MTU guard), `sock.rs` (dual-stack bind, v4-mapped send shim, #2317 outer-TOS cmsg codec, poll(2) wait layer), `attempt.rs` (#1888 S5 handshake attempt machine + keepalive emit/pace), and `dispatch.rs` (inbound type-byte dispatch with the auth-before-roam `InboundOutcome` contract + the TUN-read encap-and-send). (#6438 split; `pad_to_16` is SSOT in `afxdp/wg/mod.rs`.) |
 | `worker_manager.rs` | Per-worker lifecycle and planning state. **Two key spaces:** `live` and `identities` are keyed by binding `slot`; `records` is keyed by `worker_id`. Don't conflate them. **#6242: `records: BTreeMap<u32, WorkerRuntimeRecord>` is the SINGLE per-worker runtime registry** — it consolidates the former `handles` map plus the three sibling `Coordinator` observability maps (`worker_panics` #925 / `worker_exception_rings` / `worker_last_resolution` #5289) so a worker's handle + panic slot + exception ring + last-resolution slot register and roll back as ONE unit (`records.insert` / `records.clear`). Registration is POST-spawn-success (see `reconcile/bringup.rs`), so a failed worker never has a record to unwind (the #4952 differential is structural) and `stop_and_clear` drops all four owners of every worker in one `records.clear()`. Cold-path only: the packet path holds direct `Arc` clones and never looks up a record. Also holds `session_delta_drain_cursor` (#5290), the persistent rotating cursor for the fair RPC-fallback session-delta drain. |
 
 ## Where it sits
